split ScoreVectorSource.scala
