+++ kifi-backend/modules/search/app/com/keepit/search/engine/LibraryScoreVectorSource.scala
@@ -1,0 +1,65 @@
+package com.keepit.search.engine
+
+import com.keepit.common.akka.MonitoredAwait
+import com.keepit.search.graph.library.LibraryFields
+import com.keepit.search.{ SearchFilter, SearchConfig, Searcher }
+import com.keepit.search.engine.query.QueryProjector
+import com.keepit.search.index.WrappedSubReader
+import com.keepit.search.util.join.{ DataBuffer, DataBufferWriter }
+import org.apache.lucene.index.AtomicReaderContext
+import org.apache.lucene.search.DocIdSetIterator.NO_MORE_DOCS
+import org.apache.lucene.search.{ Query, Scorer }
+import scala.concurrent.Future
+
+class LibraryScoreVectorSource(
+    protected val searcher: Searcher,
+    protected val userId: Long,
+    protected val friendIdsFuture: Future[Set[Long]],
+    protected val libraryIdsFuture: Future[(Set[Long], Set[Long], Set[Long], Set[Long])],
+    filter: SearchFilter,
+    protected val config: SearchConfig,
+    protected val monitoredAwait: MonitoredAwait) extends ScoreVectorSourceLike with VisibilityEvaluator {
+
+  override protected def preprocess(query: Query): Query = QueryProjector.project(query, LibraryFields.textSearchFields)
+
+  protected def writeScoreVectors(readerContext: AtomicReaderContext, scorers: Array[Scorer], coreSize: Int, output: DataBuffer, directScoreContext: DirectScoreContext): Unit = {
+    val reader = readerContext.reader.asInstanceOf[WrappedSubReader]
+    val idFilter = filter.idFilter
+
+    // execute the query
+    val pq = createScorerQueue(scorers, coreSize)
+    if (pq.size <= 0) return // no scorer
+
+    val visibilityDocValues = reader.getNumericDocValues(LibraryFields.visibilityField)
+    val libraryVisibilityEvaluator = getLibraryVisibilityEvaluator(visibilityDocValues)
+
+    val idMapper = reader.getIdMapper
+    val writer: DataBufferWriter = new DataBufferWriter
+
+    val taggedScores: Array[Int] = pq.createScoreArray // tagged floats
+
+    var docId = pq.top.doc
+    while (docId < NO_MORE_DOCS) {
+      val libId = idMapper.getId(docId)
+
+      if (idFilter.findIndex(libId) < 0) { // use findIndex to avoid boxing
+        val visibility = libraryVisibilityEvaluator(docId, libId)
+
+        if (visibility != Visibility.RESTRICTED) {
+          // get all scores
+          val size = pq.getTaggedScores(taggedScores)
+
+          // write to the buffer
+          output.alloc(writer, visibility, 8 + size * 4) // id (8 bytes) and taggedFloats (size * 4 bytes)
+          writer.putLong(libId).putTaggedFloatBits(taggedScores, size)
+
+          docId = pq.top.doc // next doc
+        } else {
+          docId = pq.skipCurrentDoc() // skip this doc
+        }
+      } else {
+        docId = pq.skipCurrentDoc() // skip this doc
+      }
+    }
+  }
+}
